wldbg: allow run non-interactive server mode
If -i is not given with -s, normally load passes as if the server
mode is not active. I had to add a new option -d to optionally specify
the display, because the option -s assumed that if it is followed
by an argument, it is the display to connect to. But now there can
be passes, so change -s to be only a boolean flag and add the new
option -d.

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>